Add some CMake fixes for z/OS (not previously tested with CMake) (#795)
- Added CMake on z/OS to CI builds
- Added CMake module to enable _ALL_SOURCE if required for getrlimit(). This was already being done by Autoconf.
- Excluded pcre2_grep_test on EBCDIC platforms.  This was already being done for Autoconf builds.
- Suppressed compiler warnings for UINT16_MAX. On most platforms it's the integer constant `65535` (a signed constant), but on z/OS and possibly AIX it's defined to `65535u` (an unsigned constant), which causes warnings about unexpected signedness in comparisons and arithmetic.

+++ src/pcre2_study.c
@@ -138,7 +138,7 @@ for (;;)
   PCRE2_UCHAR op;
   PCRE2_SPTR cs, ce;
 
-  if (branchlength >= UINT16_MAX)
+  if (branchlength >= (int)UINT16_MAX)
     {
     branchlength = UINT16_MAX;
     cc = nextbranch;
@@ -627,11 +627,11 @@ for (;;)
       break;
       }
 
-     /* Take care not to overflow: (1) min and d are ints, so check that their
-     product is not greater than INT_MAX. (2) branchlength is limited to
-     UINT16_MAX (checked at the top of the loop). */
+    /* Take care not to overflow: (1) min and d are ints, so check that their
+    product is not greater than INT_MAX. (2) branchlength is limited to
+    UINT16_MAX (checked at the top of the loop). */
 
-    if ((d > 0 && (INT_MAX/d) < min) || UINT16_MAX - branchlength < min*d)
+    if ((d > 0 && (INT_MAX/d) < min) || (int)UINT16_MAX - branchlength < min*d)
       branchlength = UINT16_MAX;
     else branchlength += min * d;
     break;
@@ -2063,7 +2063,7 @@ if ((re->flags & (PCRE2_MATCH_EMPTY|PCRE2_HASACCEPT)) == 0 &&
     return 3; /* unrecognized opcode */
 
     default:
-    re->minlength = (min > UINT16_MAX)? UINT16_MAX : min;
+    re->minlength = (min > (int)UINT16_MAX)? (int)UINT16_MAX : min;
     break;
     }
   }
